dashboard

+++ app/Http/Requests/Activity/ActivityCreateRequest.php
@@ -16,7 +16,7 @@ class ActivityCreateRequest extends FormRequest
         return [
             'name'          => ['required', 'min:2', 'max:191'],
             'plan_quantity' => ['nullable', 'integer', 'min:0'],
-            'plan_time'     => ['nullable', 'numeric', 'min:0'],
+            'plan_time'     => ['nullable', 'integer', 'min:0'],
         ];
     }
 }
